[ADD] Auth server, database script and postman collection, change some configs
