+++ README.md
@@ -1,1 +1,33 @@
-# inventoryControlService
+# Inventory Control Project
+
+This repository contains the code for developing an inventory control system. The project will be implemented in multiple stages, during which additional functionalities and features will be added.
+
+## Project Stages
+
+### Stage 1: Login Service
+
+In this first stage, an authentication and authorization service will be implemented using **Spring Security** with **OAuth2**. This service will allow users to log in securely and manage their permissions within the system.
+
+## Requirements
+
+Before you start, make sure you have the following installed:
+
+- **Java 21**: The project is built using Java 21. Ensure you have the JDK installed.
+- **Maven**: This project uses Maven for dependency management. Install Maven if you haven't already.
+- **IntelliJ IDEA**: A recommended IDE for developing Java applications. You can use other IDEs, but IntelliJ provides great support for Spring projects.
+- **MySQL**: Install MySQL to manage the database for the application.
+- **Postman**: Used for testing the API endpoints.
+
+## Repository Structure
+
+The repository is organized into the following folders:
+
+- **/scripts**: Contains the scripts necessary to create the database. It includes some predefined users to facilitate testing of the system.
+
+- **/postman-collections**: Includes Postman collections that allow testing of the different APIs available in the login service. This will help developers verify the functionality of the system more efficiently.
+
+## Setup Instructions
+
+1. **Clone the repository**:
+   ```bash
+   git clone https://github.com/debigonz/inventoryControlService.git
