+++ ms-authorization-service/.gitattributes
