+++ ms-authorization-service/src/main/java/com/inventory/authorization/MsAuthorizationServiceApplication.java
@@ -1,0 +1,13 @@
+package com.inventory.authorization;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MsAuthorizationServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MsAuthorizationServiceApplication.class, args);
+    }
+
+}
