+++ ms-authorization-service/src/test/java/com/inventory/authorization/MsAuthorizationServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.inventory.authorization;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MsAuthorizationServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
