+++ ms-user-service/src/main/java/com/inventory/users/UserserviceApplication.java
@@ -1,0 +1,13 @@
+package com.inventory.users;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserserviceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserserviceApplication.class, args);
+	}
+
+}
