+++ ms-user-service/src/main/java/com/inventory/users/config/WebSecurityConfiguration.java
@@ -1,0 +1,33 @@
+
+package com.inventory.users.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+
+import static org.springframework.security.config.Customizer.withDefaults;
+
+@Configuration
+@EnableWebSecurity
+public class WebSecurityConfiguration {
+
+    @Bean
+    SecurityFilterChain securityFilterChain (HttpSecurity http) throws Exception {
+        http.authorizeHttpRequests((authorizeHttpRequests) ->
+                authorizeHttpRequests.anyRequest().authenticated())
+                .oauth2ResourceServer((oauth2ResourceServer) ->
+                        oauth2ResourceServer.jwt(withDefaults())
+                );
+        return http.build();
+    }
+
+    @Bean
+    public PasswordEncoder encoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+}
