+++ ms-user-service/src/main/java/com/inventory/users/controller/UserController.java
@@ -1,0 +1,44 @@
+package com.inventory.users.controller;
+
+import com.inventory.users.domain.entity.User;
+import com.inventory.users.service.UserService;
+import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@AllArgsConstructor
+@RestController
+@RequestMapping("/users")
+public class UserController {
+    private UserService userService;
+
+    @PostMapping("/create")
+    public ResponseEntity<User> createUser(@RequestBody User user) throws Exception {
+        return new ResponseEntity<>(userService.createUser(user), HttpStatus.CREATED);
+    }
+
+    @PutMapping("/update/{id}")
+    public ResponseEntity<User> updateUser(@PathVariable Long id) throws UsernameNotFoundException {
+        return new ResponseEntity<>(userService.updateUser(id), HttpStatus.OK);
+    }
+
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<String> deleteUser(@PathVariable Long id) throws UsernameNotFoundException {
+        userService.deleteUser(id);
+        return new ResponseEntity<>("The user with id : " + id + " was deleted", HttpStatus.OK);
+    }
+
+    @GetMapping("/user/{email}")
+    public ResponseEntity<User> getUserByEmail(@PathVariable String email) throws UsernameNotFoundException {
+        return new ResponseEntity<>(userService.findUserByEmail(email), HttpStatus.OK);
+    }
+
+    @GetMapping("/list")
+    public ResponseEntity<List<User>> getUsers() {
+        return new ResponseEntity<>(userService.findAllUsers(), HttpStatus.OK);
+    }
+}
