+++ ms-user-service/src/main/java/com/inventory/users/domain/entity/User.java
@@ -1,0 +1,19 @@
+package com.inventory.users.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Data
+@Entity
+@Table(name = "users")
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String username;
+    private String password;
+    private String email;
+    @Enumerated(EnumType.STRING)
+    private UserRole role;
+
+}
