+++ ms-user-service/src/main/java/com/inventory/users/domain/entity/UserRole.java
@@ -1,0 +1,5 @@
+package com.inventory.users.domain.entity;
+
+public enum UserRole {
+    ADMIN,USER
+}
