+++ ms-user-service/src/main/java/com/inventory/users/domain/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.inventory.users.domain.repository;
+
+import com.inventory.users.domain.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
