+++ ms-user-service/src/main/resources/application.yml
@@ -1,0 +1,24 @@
+spring:
+  application:
+    name: ms-users
+  security:
+    oauth2:
+      resource server:
+        jwt:
+          issuer-uri: http://localhost:8081
+  datasource:
+    url: jdbc:mysql://localhost:3306/user_database?useSSL=false&serverTimezone=UTC
+    username: root
+    password: root
+  jpa:
+    properties:
+      hibernate.dialect: org.hibernate.dialect.MySQLDialect
+    database: MYSQL
+
+server:
+  port: 8080
+
+
+logging:
+  level:
+    org: DEBUG
