+++ ms-user-service/src/test/java/com/inventory/users/UserserviceApplicationTests.java
@@ -1,0 +1,14 @@
+package com.inventory.users;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserserviceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+
+}
