--- userservice/src/main/java/com/inventory/userservice/UserserviceApplication.java
@@ -1,13 +1,0 @@
-package com.inventory.userservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserserviceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserserviceApplication.class, args);
-	}
-
-}
