--- userservice/src/main/java/com/inventory/userservice/config/WebSecurityConfiguration.java
@@ -1,34 +1,0 @@
-/*
-package com.inventory.userservice.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.oauth2.jwt.JwtDecoder;
-import org.springframework.security.oauth2.jwt.JwtDecoders;
-import org.springframework.security.web.SecurityFilterChain;
-
-import static org.springframework.security.config.Customizer.withDefaults;
-
-@Configuration
-@EnableWebSecurity
-public class WebSecurityConfiguration {
-
-    @Bean
-    SecurityFilterChain defaultSecurityFilterChain (HttpSecurity http) throws Exception {
-        http.authorizeHttpRequests((authorizeHttpRequests) ->
-                authorizeHttpRequests.anyRequest().authenticated())
-                .oauth2ResourceServer((oauth2ResourceServer) ->
-                        oauth2ResourceServer.jwt(withDefaults())
-                );
-        return http.build();
-    }
-
-    @Bean
-    public JwtDecoder jwtDecoder() {
-        return JwtDecoders.fromIssuerLocation("https://accounts.google.com");
-    }
-
-}
-*/
