--- userservice/src/main/java/com/inventory/userservice/controller/UserController.java
@@ -1,44 +1,0 @@
-package com.inventory.userservice.controller;
-
-import com.inventory.userservice.domain.entity.User;
-import com.inventory.userservice.service.UserService;
-import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-//import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@AllArgsConstructor
-@RestController( "/users")
-public class UserController {
-    private UserService userService;
-
-    @PostMapping("/create")
-    public ResponseEntity<User> createUser(@RequestBody User user) throws Exception {
-        return new ResponseEntity<>(userService.createUser(user), HttpStatus.CREATED);
-    }
-
-    @PutMapping("/update")
-    public ResponseEntity<User> updateUser(@RequestBody User user) throws UsernameNotFoundException {
-        return new ResponseEntity<>(userService.updateUser(user), HttpStatus.OK);
-    }
-
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable Long id) throws UsernameNotFoundException {
-        userService.deleteUser(id);
-        return new ResponseEntity<>("The user : " + id + " was deleted", HttpStatus.OK);
-    }
-
-    @GetMapping("/user/{email}")
-    public ResponseEntity<User> getUserByEmail(@PathVariable String email) throws UsernameNotFoundException {
-        return new ResponseEntity<>(userService.findUserByEmail(email), HttpStatus.OK);
-    }
-
-    @GetMapping("/list")
-    public ResponseEntity<List<User>> getUsers() {
-        return new ResponseEntity<>(userService.findAllUsers(), HttpStatus.OK);
-    }
-}
