--- userservice/src/main/java/com/inventory/userservice/domain/entity/UserRole.java
@@ -1,5 +1,0 @@
-package com.inventory.userservice.domain.entity;
-
-public enum UserRole {
-    ADMIN,USER
-}
