--- userservice/src/main/java/com/inventory/userservice/domain/model/UserDTO.java
@@ -1,14 +1,0 @@
-package com.inventory.userservice.domain.model;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class UserDTO {
-    private Long id;
-    private String username;
-    private String password;
-    private String email;
-    private UserRoleDTO role;
-}
