--- userservice/src/main/java/com/inventory/userservice/domain/model/UserRoleDTO.java
@@ -1,5 +1,0 @@
-package com.inventory.userservice.domain.model;
-
-public enum UserRoleDTO {
-    ADMIN, USER
-}
