--- userservice/src/main/java/com/inventory/userservice/domain/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.inventory.userservice.domain.repository;
-
-import com.inventory.userservice.domain.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
