--- userservice/src/main/java/com/inventory/userservice/service/UserService.java
@@ -1,56 +1,0 @@
-package com.inventory.userservice.service;
-
-import com.inventory.userservice.domain.entity.User;
-import com.inventory.userservice.domain.repository.UserRepository;
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Slf4j
-@AllArgsConstructor
-@Service
-public class UserService implements UserDetailsService {
-    private UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException(email));
-    }
-
-    public User createUser(User user) throws Exception {
-        boolean userExists = userRepository.findByEmail(user.getEmail()).isPresent();
-        if (userExists) {
-            throw new Exception("User already exists");
-        }
-        return userRepository.save(user);
-    }
-
-    public User updateUser(User user) throws UsernameNotFoundException {
-        boolean userExists = userRepository.findByEmail(user.getEmail()).isPresent();
-        if (userExists) {
-            userRepository.save(user);
-        } else {
-            throw new UsernameNotFoundException(user.getEmail());
-        }
-        return user;
-    }
-
-    public void deleteUser(Long id) {
-        User user = userRepository.findById(id).orElseThrow(() -> new UsernameNotFoundException(id.toString()));
-        userRepository.delete(user);
-    }
-
-    public User findUserByEmail(String email) {
-        return userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException(email));
-    }
-
-    public List<User> findAllUsers() {
-        return userRepository.findAll();
-    }
-
-}
