--- userservice/src/main/resources/application.yml
@@ -1,29 +1,0 @@
-spring:
-  application:
-    name: user-service
-  security:
-    oauth2:
-      resource server:
-        jwt:
-          issuer-uri: https://accounts.google.com
-  datasource:
-    driverClassName: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/user_database
-    username: root
-    password: root
-    sql:
-      init:
-        mode: always
-        dataLocations: classpath:/user-service.sql
-    jpa:
-      deferDatasourceInitialization: true
-      hibernate:
-        ddlAuto: create-drop
-
-  server:
-    port: 8080
-
-
-logging:
-  level:
-    org: DEBUG
