--- userservice/src/main/resources/user-service.sql
@@ -1,1 +1,0 @@
-INSERT INTO users(username, password, email, role) VALUES('debigonz', 'Test1234', 'debigonz19@gmail.com', 'ADMIN');
