--- userservice/src/test/java/com/inventory/userservice/UserserviceApplicationTests.java
@@ -1,15 +1,0 @@
-package com.inventory.userservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserserviceApplicationTests {
-/*
-	@Test
-	void contextLoads() {
-	}
-
-
- */
-}
